The most significant changes involve the addition of a new parameter `CloseApplication` of type `EventCallback` to the `CustomIntelliSenseTable` and `CustomIntelliSense` classes. This parameter is used to trigger the closing of the application. It replaces previous code that was used to close the application and is passed to the respective components in the `Index.razor` file.
In the `Index.razor.cs` file, a new method `CloseWindow` has been added which invokes the `Callback` event. A new `Callback` parameter of type `EventCallback` has also been added to this class.

In the `MainForm.cs` file, the `button1_Click` method that was previously used to close the application has been removed. Instead, the `Callback` event is passed to the `Index` component when it is added to the `blazorWebView1` object, and this event is set to close the application.

In the `Program.cs` file, an event handler has been added for the `UnhandledException` event of the `AppDomain.CurrentDomain` object. This event handler shows a message box with the error information when an unhandled exception occurs.

Finally, in the `index.html` file, a new JavaScript function `closeForm` has been added. This function invokes a .NET method `CloseForm` in the `WinFormsApp` namespace.

List of changes:

1. Addition of `CloseApplication` parameter to `CustomIntelliSenseTable` and `CustomIntelliSense` classes.
2. Invocation of `CloseApplication` parameter in several places in the aforementioned classes.
3. Passing of `CloseApplication` parameter to the components in `Index.razor` file.
4. Addition of `CloseWindow` method and `Callback` parameter in `Index.razor.cs` file.
5. Removal of a button used to close the application in `MainForm.Designer.cs` file.
6. Removal of `button1_Click` method and addition of `Callback` event in `MainForm.cs` file.
7. Addition of an event handler for `UnhandledException` event in `Program.cs` file.
8. Addition of `closeForm` JavaScript function in `index.html` file.