feat: Add JsonRepositoryOptions class and ICursorlessCheatsheetItemJsonRepository interface
The code changes include adding a new class `JsonRepositoryOptions` in the `DataAccessLibrary\Repositories` namespace and an interface `ICursorlessCheatsheetItemJsonRepository` in the `DataAccessLibrary\Repositories` namespace. The `JsonRepositoryOptions` class has a property `JsonFilePath` of type string. The `ICursorlessCheatsheetItemJsonRepository` interface defines several methods for retrieving and searching cursorless cheatsheet items asynchronously, as well as exporting to JSON.

Recent user commits and repository commits do not provide any additional context for the commit message.