feat: Add SelectWordFromRight property to CustomIntelliSenseDTO and CustomIntelliSense model
The code changes include adding a new property, SelectWordFromRight, to the CustomIntelliSenseDTO and CustomIntelliSense models. This property allows users to specify the number of words to select from the right when using the IntelliSense feature. This feature enhances the user experience by providing more flexibility in selecting words for auto-completion.